feat(common, ..): move `Path` to dedicated file
