+++ src/plugins/common/src/tools.rs
@@ -10,6 +10,7 @@ pub mod item_type;
 pub mod iter_helpers;
 pub mod movement_animation;
 pub mod ordered_hash_map;
+pub mod path;
 pub mod skill_execution;
 pub mod speed;
 pub mod swap_key;
